feat: 分离基础服务启动命令

+++ deploy/docker-compose/deploy.sh
@@ -11,11 +11,27 @@ export INFLUXDB_PASSWORD="password"
 export REDIS_PASSWORD="123456"
 
 if [[ $1 = "up" ]]; then
-  docker-compose -f base_service.yaml up -d
-  docker-compose -f server.yaml up -d
+  if [[ $2 = "" ]]; then
+    docker-compose -f base_service.yaml up -d
+    docker-compose -f server.yaml up -d
+  fi
+  if [[ $2 = "base" ]]; then
+    docker-compose -f base_service.yaml up -d
+  fi
+  if [[ $2 = "dev" ]]; then
+    docker-compose -f server.yaml up -d
+  fi
 fi
 
 if [[ $1 = "down" ]]; then
-  docker-compose -f base_service.yaml down
-  docker-compose -f server.yaml down
+  if [[ $2 = "" ]]; then
+    docker-compose -f base_service.yaml down
+    docker-compose -f server.yaml down
+  fi
+  if [[ $2 = "base" ]]; then
+    docker-compose -f base_service.yaml down
+  fi
+  if [[ $2 = "dev" ]]; then
+    docker-compose -f server.yaml down
+  fi
 fi
